upadte: Fixing the type for the Availability zone for the NNC to uint type

+++ crd/nodenetworkconfig/api/v1alpha/nodenetworkconfig.go
@@ -50,7 +50,7 @@ type NodeNetworkConfigSpec struct {
 	// AvailabilityZone contains the Azure availability zone for the virtual machine where network containers are placed.
 	// NMA returns an int value for the availability zone.
 	// +kubebuilder:validation:Optional
-	AvailabilityZone int64 `json:"availabilityZone,omitempty"`
+	AvailabilityZone uint `json:"availabilityZone,omitempty"`
 }
 
 // Status indicates the NNC reconcile status
